server/requirements.txt
# requirements.txt
Flask
Flask-Cors
google-generativeai
firebase-admin
python-dotenv
gunicorn # For production deployment